Added extended nagios monitoring (raid/smart/etc)
This installs nrpe which is used for nagios monitoring to execute
pre-set commands so you can do monitoring of things that can't be
done passively (like http/ping) such as uptime, smart, raid, etc.

Also included are some perl scripts for getting raid status and
also smart info off SATA disks. SAS disks are not currently being
looked at as smart is less important and they are less likely to
cause issues when the drives do start failing but will be added
at another time.

Also included are the cli tools for megacli and areca's cli64 as
it is likely we will have areca machines in the mix in the future.

Signed-off-by: Sandon Van Ness <sandon@inktank.com>
Reviewed-by: Dan Mick <dan.mick@inktank.com>