raid.pl: Fix raid detection for JBOD mode; smart.pl: Crit for firm.
Fixed raid.pl to look at the whole line instead of splitting the
words up as some controllers in JBOD mode show different output
so this method should be more reliable. Also added critical for
old areca firmware for smart checking so it will show up on nagios
as an error/critical instead of 'ok' when the machine has older
firmware on it.