Removed networking restart
This is not really needed now that most machines have had
ceph-qa-chef ran on them. Also it runs on them during the end of
the preseed during imaging and they reboot right after that so
there is really no need for it there either. It was only needed
in the first place to deal with machines that did not yet have their
interfaces file updated to have static IPs to take effect immediately.

Signed-off-by: Sandon Van Ness <sandon@inktank.com>