For VM's switch source.list to use apt-mirror server.
In order to prevent possible apt failures be sure to change
/etc/apt/sources.list to use apt-mirror.front.sepia.ceph.com
as one of the first things it does. This is mainly for vms that
get spun up. Since they are ubuntu's images we want to set this
before any apt commands are ran to insure it doesn't get apt
failures.

Signed-off-by: Sandon Van Ness <sandon@inktank.com>