Mime-type preinstall pre 2.0 to avoid chef installation failure.
RHEL/CentOS does not have ruby 1.9.2 which is required by
mime-type which is attempted to be installed as a dependency to
chef. Chef probably wont need mime-type 2.0 for a long time but
the gem dependency system doesn't seem smart enough to use the
older version automatically. This pre-installs mim-type 1.25
which is the newest pre 2.0 version so doing a gem install chef
suceeds.

Signed-off-by: Sandon Van Ness <sandon@inktank.com>